Extract the composition root into includes/class-ss-shipping-wc.php
smart-send-logistics.php shrinks to the thin plugin entry file (55 lines):
the plugin header, the SS_SHIPPING_PLUGIN_FILE constant, one require, the
SS_SHIPPING_WC() accessor and the bootstrap instantiation. The
SS_Shipping_WC class moves (git mv, so history follows) to
includes/class-ss-shipping-wc.php, reading SS_SHIPPING_PLUGIN_FILE where
it historically used __FILE__ - the only content change the split needs.
phpcs baseline paths follow the move.

Part of #140.
